+++ src/Linna/Authentication/ProtectedController.php
@@ -11,6 +11,8 @@ declare(strict_types=1);
 
 namespace Linna\Authentication;
 
+use Linna\Authentication\Exception\AuthenticationException;
+
 /**
  * Help protect a controller with login.
  */
@@ -22,16 +24,32 @@ trait ProtectedController
     private $authentication = false;
 
     /**
-     * Allow access to controller only if logged.
+     * Allow access to controller class or methods only if logged.
+     * Return a status code, useful with AJAX requests.
      *
      * @param Authentication $authentication
-     * @param string         $redirect
+     * @param int            $httpResponseCode
      */
-    private function protect(Authentication $authentication, string $redirect): void
+    private function protect(Authentication $authentication, int $httpResponseCode = 403): void
     {
         if (($this->authentication = $authentication->isLogged()) === false) {
-            header('Location: '.$redirect);
-            exit;
+            http_response_code($httpResponseCode);
+            throw new AuthenticationException('');
+        }
+    }
+
+    /**
+     * Allow access to controller class or methods only if logged
+     * and do a redirection.
+     *
+     * @param Authentication $authentication
+     * @param string         $location
+     */
+    private function protectWithRedirect(Authentication $authentication, string $location): void
+    {
+        if (($this->authentication = $authentication->isLogged()) === false) {
+            header('Location: '.$location);
+            throw new AuthenticationException('');
         }
     }
 }